requirements_development.txt
-r ./requirements.txt
django-debug-toolbar
mypy
django-stubs[compatible-mypy]
